change page to resource, add type-map
Changes all references to page to resource.  Also add a type-map and an
additional form of defresource that accepts a type argument and stores
it in the type-map.  Some error-checking is also added to the
add-bb-entry function that makes sure the message contains both keys,
and only uses both keys.