Adds support for adding an image
This commit adds support for sending images, or any other kind of binary
data.  Bit of  big change made...resource functions now return a byte
array.  There was a problem with sending binary read from a file as a
string.  It would be cleaner to add a function that coerces data into a
byte array instead of expecting it.

There is also a change made to the response returned by gen-response.
It now returns a map structure in a similar manner to extract-req.

Add predicate function accept-gzip? that tests if the user agent will
accept gzip encoding as a response.

Remove :post and :head method functions until new changes are ironed
out.

Replace write-to-socket with different function write-response-to-stream
that write the mapped response data to a stream, using new function
write-string-to-stream.