Move bulletin board code into its own namespace
adds namespace lehs.demo.bb for the bulletin board.  Also improves upon
the look of the bb table (via an update to foo.css) and alternating the
table row class field to give it alternating dark/light colored rows.
Finally, sets the location header of the response to /d#<objectid>.