Refactor, adds concurrency & kill /killserver
Adds function kill-server, which sets the value of the reference
kill-server? to true.  Change name of
accept-connection-and-send-response to read-req-and-send-response.
Refactors the function run-server into three different functions.

run-server just opens the server socket and calls the server-loop
function.

server-loop awaits connections; it times out every so often to check if
the kill-server? flag has been set; if so, it returns.  It calls the
accept-connection function to accept connections.

accept-connection actually calls .accept on the server socket.  If a
client connects, the socket returned socket is passed off to a future to
handle the connection in read-req-and-send-response.